+++ Lamune_101878/MNK.lua
@@ -73,15 +73,19 @@ local sets = {
         Feet = 'Herald\'s Gaiters',
     },
 
+    --[[
+    8% Base
+    5% Merits
+    45% Counterstance
+    10% Melee Gaiters
+    ]]
     DT = {
         Main = 'Cross-Counters', -- 5
         Ammo = 'Fenrir\'s Stone',
         Head = 'Optical Hat',
         Neck = 'Faith Torque',
         Ear1 = 'Brutal Earring',
-        Ear2 = 'Merman\'s Earring',
-        -- Ear1 = 'Avenger\'s Earring', -- 1
-        -- Ear2 = 'Avenger\'s Earring', -- 1
+        Ear2 = 'Avenger\'s Earring', -- 1
         Body = 'Scp. Harness +1',
         Hands = 'Mel. Gloves +1',
         -- Hands = 'Rasetsu Tekko +1', -- 1
@@ -285,6 +289,8 @@ Everything below can be ignored.
 gcmelee = gFunc.LoadFile('common\\gcmelee.lua')
 
 profile.HandleAbility = function()
+    gcmelee.DoAbility()
+
     local action = gData.GetAction()
 
     if string.match(action.Name, 'Jump') then