+++ Lamune_101878/SMN.lua
@@ -45,7 +45,7 @@ local sets = {
         Hands = 'Hydra Gloves',
         Ring1 = 'Bomb Queen Ring',
         Ring2 = 'Ether Ring',
-        Back = 'Errant Cape',
+        Back = 'Mahatma Cape',
         Waist = 'Hierarch Belt',
         Legs = 'Hydra Brais',
         Feet = 'Hydra Gaiters',
@@ -181,7 +181,7 @@ local sets = {
         Hands = 'Merman\'s Bangles',
         Ring1 = 'Sattva Ring',
         Ring2 = 'Evoker\'s Ring',
-        Back = { Name = 'Errant Cape', Priority = 100 },
+        Back = { Name = 'Mahatma Cape', Priority = 100 },
         Waist = 'Silver Obi +1', -- 8
         Legs = { Name = 'Evk. Spats +1', Priority = 100 },
         Feet = { Name = 'Mountain Gaiters', Priority = 100 }, -- 5
@@ -198,7 +198,7 @@ local sets = {
         Hands = 'Merman\'s Bangles',
         Ring1 = 'Sattva Ring',
         Ring2 = 'Evoker\'s Ring',
-        Back = { Name = 'Errant Cape', Priority = 100 },
+        Back = { Name = 'Mahatma Cape', Priority = 100 },
         Waist = 'Silver Obi +1', -- 8
         Legs = { Name = 'Evk. Spats +1', Priority = 100 },
         Feet = { Name = 'Mountain Gaiters', Priority = 100 }, -- 5
@@ -227,7 +227,7 @@ local sets = {
         Hands = 'Hydra Gloves', -- 5
         Ring1 = 'Aqua Ring',
         Ring2 = 'Communion Ring',
-        Back = { Name = 'Errant Cape', Priority = 100 }, -- 5
+        Back = { Name = 'Mahatma Cape', Priority = 100 }, -- 5
         Waist = 'Penitent\'s Rope', -- 3
         Legs = 'Hydra Brais', -- 6
         Feet = 'Hydra Gaiters', -- 5
@@ -373,6 +373,7 @@ gcmage = gFunc.LoadFile('common\\gcmage.lua')
 
 profile.HandleAbility = function()
     gcmage.DoAbility()
+
     gFunc.EquipSet('BP_Delay')
 end