+++ Rag_5040/BRD.lua
@@ -8,8 +8,6 @@ local whmSJMaxMP = 188 -- The Max MP you have when /whm in your idle set
 local rdmSJMaxMP = nil -- The Max MP you have when /rdm in your idle set
 local blmSJMaxMP = nil -- The Max MP you have when /blm in your idle set
 
-local displayheadOnAbility = true
-
 local sets = {
     Idle = {
         Main = 'Terra\'s Staff',
@@ -24,7 +22,7 @@ local sets = {
         Ring1 = 'Shadow Ring',
         Ring2 = 'Sattva Ring',
         Back = 'Umbra Cape',
-        Waist = { Name = 'Hierarch Belt', Priority = 100 },
+        Waist = { Name = 'Ocean Rope', Priority = 100 },
         Legs = 'Dst. Subligar +1',
         Feet = 'Dst. Leggings +1',
     },
@@ -41,7 +39,7 @@ local sets = {
         Ring1 = 'Shadow Ring',
         Ring2 = 'Sattva Ring',
         Back = 'Umbra Cape',
-        Waist = { Name = 'Hierarch Belt', Priority = 100 },
+        Waist = { Name = 'Ocean Rope', Priority = 100 },
         Legs = 'Dst. Subligar +1',
         Feet = 'Dst. Leggings +1',
     },
@@ -57,8 +55,8 @@ local sets = {
         Hands = 'Hydra Gloves',
         Ring1 = "Shadow Ring",
         Ring2 = "Merman's Ring",
-        Back = 'Errant Cape',
-        Waist = { Name = 'Hierarch Belt', Priority = 100 },
+        Back = 'Mahatma Cape',
+        Waist = { Name = 'Ocean Rope', Priority = 100 },
         Legs = 'Hydra Brais',
         Feet = 'Hydra Gaiters',
     },
@@ -86,7 +84,7 @@ local sets = {
         Ring1 = 'Shadow Ring',
         Ring2 = 'Sattva Ring', -- 5
         Back = 'Hexerei Cape', -- 3
-        Waist = { Name = 'Hierarch Belt', Priority = 100 },
+        Waist = { Name = 'Ocean Rope', Priority = 100 },
         Legs = 'Byakko\'s Haidate',
         Feet = 'Suzaku\'s Sune-ate',
     },
@@ -208,7 +206,7 @@ local sets = {
     Sing_Buff = {
         Main = 'Chanter\'s Staff',
         Body = 'Minstrel\'s Coat',
-        Waist = { Name = 'Hierarch Belt', Priority = 100 },
+        Waist = { Name = 'Ocean Rope', Priority = 100 },
         Feet = 'Dusk Ledelsens +1',
     },
     Sing_Debuff = {
@@ -234,7 +232,7 @@ local sets = {
         -- Range = 'Harlequin\'s Horn',
     },
     Sing_Minuet = {
-        Range = 'Gjallarhorn',
+        Range = 'Cornette +2',
     },
     Sing_March = {
         Range = 'Faerie Piccolo',
@@ -271,7 +269,7 @@ local sets = {
         Legs = "Mahatma Slops",
     },
     Sing_Requiem = {
-        Range = 'Gjallarhorn',
+        Range = 'Hamelin Flute',
     },
     Sing_Carol = {
         Range = 'Crumhorn +1',
@@ -315,7 +313,7 @@ local sets = {
         Hands = 'Hydra Gloves', -- 5
         Ring1 = 'Aqua Ring',
         Ring2 = 'Communion Ring',
-        Back = { Name = 'Errant Cape', Priority = 100 }, -- 5
+        Back = { Name = 'Mahatma Cape', Priority = 100 }, -- 5
         Waist = 'Penitent\'s Rope', -- 3
         -- Legs = 'Mahatma Slops', -- 4
         Legs = 'Hydra Brais', -- 6
@@ -386,9 +384,7 @@ Everything below can be ignored.
 gcmage = gFunc.LoadFile('common\\gcmage.lua')
 
 profile.HandleAbility = function()
-    if (displayheadOnAbility) then
-        AshitaCore:GetChatManager():QueueCommand(-1, '/displayhead')
-    end
+    gcmage.DoAbility()
 end
 
 profile.HandleItem = function()