+++ common/gcmage.lua
@@ -48,6 +48,9 @@ local healers_earring_slot = 'Ear2'
 local tp_fencers_ring = false
 local tp_fencers_ring_slot = ''
 
+-- BLM / SMN Specific
+local claustrum = false
+
 -- BLM Specific
 local sorcerers_ring = false
 local sorcerers_ring_slot = '' -- This is Ring1 instead of Ring2 to allow Ice Ring override to work
@@ -366,6 +369,9 @@ function gcmage.DoDefault(ninSJMMP, whmSJMMP, blmSJMMP, rdmSJMMP, drkSJMMP)
             local staff = ElementalStaffTable[lastSummoningElement]
             if (bahamuts_staff) then
                 staff = 'Bahamut\'s Staff'
+            end
+            if (claustrum) then
+                staff = 'Claustrum'
             end
             if staff ~= '' then
                 gFunc.Equip('Main', staff)
@@ -918,6 +924,11 @@ function gcmage.EquipStaff()
 
     if (action.Skill ~= 'Enhancing Magic' and not ElementalDebuffs:contains(action.Name) and not string.match(action.Name, 'Utsusemi')) then
         local staff = ElementalStaffTable[action.Element]
+        if (player.MainJob == 'SMN' or player.MainJob == 'BLM') then
+            if (claustrum and action.Skill ~= 'Healing Magic') then
+                staff = 'Claustrum'
+            end
+        end
         if staff ~= '' then
             gFunc.Equip('Main', staff)
         end
@@ -957,6 +968,7 @@ function ObiCheck(action)
 end
 
 function gcmage.DoAbility()
+    gcinclude.DoAbility()
     local action = gData.GetAction()
     
     if (action.Name == 'Release') then